🚚 move `mergeInto` and `typeUtils` utilities to `@datadog/js-core/utils`
`combine`, `deepClone`, `mergeInto`, `getType`, and `isIndexableObject`
are not browser-specific, so they belong in the shared `js-core` package
rather than `browser-core`. This makes them reusable by future non-browser
targets without pulling in the full `browser-core` dependency.
