+++ packages/browser-core/src/domain/telemetry/telemetry.ts
@@ -1,5 +1,5 @@
 import { clocksNow } from '@datadog/js-core/time'
-import { getDebugMode } from '@datadog/js-core/util'
+import { getDebugMode, combine } from '@datadog/js-core/util'
 import type { Context } from '../../tools/serialisation/context'
 import { NO_ERROR_STACK_PRESENT_MESSAGE, isError } from '../error/error'
 import { toStackTraceString } from '../../tools/stackTrace/handlingStack'
@@ -14,7 +14,6 @@ import { display } from '../../tools/display'
 import { sendToExtension } from '../../tools/sendToExtension'
 import { performDraw } from '../../tools/utils/numberUtils'
 import { jsonStringify } from '../../tools/serialisation/jsonStringify'
-import { combine } from '../../tools/mergeInto'
 import { NonErrorPrefix } from '../error/error.types'
 import type { StackTrace } from '../../tools/stackTrace/computeStackTrace'
 import { computeStackTrace } from '../../tools/stackTrace/computeStackTrace'
