+++ packages/browser-core/src/index.ts
@@ -172,11 +172,9 @@ export * from './tools/utils/objectUtils'
 export * from './tools/utils/functionUtils'
 export * from './tools/serialisation/jsonStringify'
 export * from './tools/serialisation/stringify'
-export * from './tools/mergeInto'
 export * from './tools/utils/stringUtils'
 export * from './tools/matchOption'
 export * from './tools/utils/responseUtils'
-export * from './tools/utils/typeUtils'
 export type { RawError, RawErrorCause, ErrorWithCause, Csp } from './domain/error/error.types'
 export { ErrorHandling, ErrorSource } from './domain/error/error.types'
 export * from './domain/deflate'
