+++ packages/browser-core/src/tools/utils/browserDetection.spec.ts
@@ -1,4 +1,4 @@
-import { combine } from '../mergeInto'
+import { combine } from '@datadog/js-core/util'
 import { Browser, detectBrowser } from './browserDetection'
 
 describe('browserDetection', () => {
