+++ packages/browser-logs/src/domain/logger/loggerCollection.ts
@@ -1,8 +1,8 @@
 import { timeStampNow } from '@datadog/js-core/time'
 import type { TimeStamp } from '@datadog/js-core/time'
 import type { Context } from '@datadog/browser-core'
-import { combine, ErrorSource } from '@datadog/browser-core'
-import { originalConsoleMethods, globalConsole, ConsoleApiName } from '@datadog/js-core/util'
+import { ErrorSource } from '@datadog/browser-core'
+import { originalConsoleMethods, globalConsole, ConsoleApiName, combine } from '@datadog/js-core/util'
 import type { CommonContext, RawLogsEvent } from '../../rawLogsEvent.types'
 import type { LifeCycle, RawLogsEventCollectedData } from '../lifeCycle'
 import { LifeCycleEventType } from '../lifeCycle'
