+++ packages/browser-rum-core/src/domain/error/errorCollection.ts
@@ -1,5 +1,5 @@
 import type { ClocksState } from '@datadog/js-core/time'
-import { ConsoleApiName } from '@datadog/js-core/util'
+import { ConsoleApiName, combine } from '@datadog/js-core/util'
 import type { Context, RawError, BufferedData } from '@datadog/browser-core'
 import {
   BufferedDataType,
@@ -9,7 +9,6 @@ import {
   computeRawError,
   ErrorHandling,
   NonErrorPrefix,
-  combine,
 } from '@datadog/browser-core'
 import type { RawRumErrorEvent } from '../../rawRumEvent.types'
 import { RumEventType } from '../../rawRumEvent.types'
