+++ packages/browser-rum-core/src/domain/eventTracker.ts
@@ -1,7 +1,8 @@
 import type { ClocksState, RelativeTime, Duration } from '@datadog/js-core/time'
 import type { ValueHistoryEntry } from '@datadog/browser-core'
 import { ONE_MINUTE, elapsed } from '@datadog/js-core/time'
-import { generateUUID, createValueHistory, combine } from '@datadog/browser-core'
+import { generateUUID, createValueHistory } from '@datadog/browser-core'
+import { combine } from '@datadog/js-core/util'
 import type { RumActionEvent, RumErrorEvent, RumLongTaskEvent, RumResourceEvent } from '../rumEvent.types'
 import { LifeCycleEventType } from './lifeCycle'
 import type { LifeCycle } from './lifeCycle'
