+++ packages/browser-rum-core/src/domain/limitModification.ts
@@ -1,5 +1,6 @@
-import { sanitize, deepClone, getType, objectEntries } from '@datadog/browser-core'
+import { sanitize, objectEntries } from '@datadog/browser-core'
 import type { Context } from '@datadog/browser-core'
+import { deepClone, getType } from '@datadog/js-core/util'
 
 export type ModifiableFieldPaths = Record<string, 'string' | 'object'>
