+++ packages/browser-rum-core/src/domain/resource/resourceCollection.ts
@@ -1,7 +1,6 @@
 import type { Duration } from '@datadog/js-core/time'
 import { toServerDuration, relativeToClocks } from '@datadog/js-core/time'
 import {
-  combine,
   generateUUID,
   createTaskQueue,
   mockable,
@@ -13,6 +12,7 @@ import {
   RequestType,
   setTimeout,
 } from '@datadog/browser-core'
+import { combine } from '@datadog/js-core/util'
 import type { MatchHeader, RumConfiguration } from '../configuration'
 import { RumPerformanceEntryType, createPerformanceObservable } from '../../browser/performanceObservable'
 import type { RumResourceEventDomainContext } from '../../domainContext.types'
