+++ packages/browser-rum-core/test/fixtures.ts
@@ -1,7 +1,8 @@
 import type { RelativeTime, Duration, ServerDuration, TimeStamp } from '@datadog/js-core/time'
 import type { Context } from '@datadog/browser-core'
 import { relativeNow } from '@datadog/js-core/time'
-import { combine, ErrorHandling, ErrorSource, generateUUID, ResourceType } from '@datadog/browser-core'
+import { ErrorHandling, ErrorSource, generateUUID, ResourceType } from '@datadog/browser-core'
+import { combine } from '@datadog/js-core/util'
 import type { RumMutationRecord } from '@datadog/browser-rum-core'
 import {
   type RumPerformanceEntry,
