+++ packages/js-core/src/entries/util.ts
@@ -1,3 +1,5 @@
 export { createDisplay, ConsoleApiName, globalConsole, originalConsoleMethods } from '../util/display'
 export type { Display } from '../util/display'
 export { setDebugMode, getDebugMode } from '../util/debug'
+export * from '../util/mergeInto'
+export * from '../util/typeUtils'
