+++ packages/js-core/src/util/mergeInto.spec.ts
